Добавление/удаление роли (back)

+++ backend/src/main/java/com/backend/bprobackend/repository/RoleRepos.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepos extends JpaRepository<Role,Long> {
+public interface RoleRepos extends JpaRepository<Role,Integer> {
     Optional<Role> findByName(EnumRole name);
 
 }
